+++ gradle.properties
@@ -54,7 +54,7 @@ accessTransformersFile =
 # Provides setup for Mixins if enabled. If you don't know what mixins are: Keep it disabled!
 usesMixins = true
 # Specify the package that contains all of your Mixins. You may only place Mixins in this package or the build will fail!
-mixinsPackage = mixins
+mixinsPackage = mixins.impl
 # Specify the core mod entry class if you use a core mod. This class must implement IFMLLoadingPlugin!
 # Example value: coreModClass = asm.FMLPlugin + modGroup = com.myname.mymodid -> com.myname.mymodid.asm.FMLPlugin
 coreModClass = GTEMixinPlugin
