+++ src/main/java/gtexpert/mixins/impl/ezstorage2/MixinEZInventory.java
@@ -1,4 +1,4 @@
-package gtexpert.mixins;
+package gtexpert.mixins.impl.ezstorage2;
 
 import net.minecraft.item.ItemStack;
